:recycle: OAuthClientの定義を移動 #22
