+++ internal/model/oauth.go
@@ -3,55 +3,15 @@ package model
 import (
 	"context"
 	"fmt"
-	"os"
 	"time"
 
 	"golang.org/x/oauth2"
-	"google.golang.org/api/calendar/v3"
-	goauth2 "google.golang.org/api/oauth2/v2"
-	"google.golang.org/api/option"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
 type OAuth struct {
 	Config *oauth2.Config
-}
-
-func NewGoogleOAuth() *OAuth {
-	cfg := &oauth2.Config{
-		ClientID:     os.Getenv("GOOGLE_CLIENT_ID"),
-		ClientSecret: os.Getenv("GOOGLE_CLIENT_SECRET"),
-		Endpoint: oauth2.Endpoint{
-			AuthURL:  os.Getenv("GOOGLE_AUTH_URL"),
-			TokenURL: os.Getenv("GOOGLE_TOKEN_URL"),
-		},
-		RedirectURL: os.Getenv("GOOGLE_REDIRECT_URL"),
-		Scopes: []string{
-			goauth2.UserinfoProfileScope,
-			calendar.CalendarReadonlyScope,
-		},
-	}
-	oauth := &OAuth{
-		Config: cfg,
-	}
-	return oauth
-}
-
-func NewNotionOAuth() *OAuth {
-	cfg := &oauth2.Config{
-		ClientID:     os.Getenv("NOTION_CLIENT_ID"),
-		ClientSecret: os.Getenv("NOTION_CLIENT_SECRET"),
-		Endpoint: oauth2.Endpoint{
-			AuthURL:  os.Getenv("NOTION_AUTH_URL"),
-			TokenURL: os.Getenv("NOTION_TOKEN_URL"),
-		},
-		RedirectURL: os.Getenv("NOTION_REDIRECT_URL"),
-	}
-	oauth := &OAuth{
-		Config: cfg,
-	}
-	return oauth
 }
 
 func (o *OAuth) GetAuthCodeURL(oauthState string) string {
@@ -74,21 +34,6 @@ func (o *OAuth) RefreshToken(ctx context.Context, userID string) (*oauth2.Token,
 		return nil, err
 	}
 	return newToken, nil
-}
-
-// 別のModelに書きべきかも
-// Why: oauth2とは別のUserinfoAPIを叩くため
-func (o *OAuth) GetUserID(ctx context.Context, token *oauth2.Token) (string, error) {
-	os, err := goauth2.NewService(ctx, option.WithTokenSource(o.Config.TokenSource(ctx, token)))
-	if err != nil {
-		return "", err
-	}
-	us := goauth2.NewUserinfoService(os)
-	userinfo, err := us.Get().Do()
-	if err != nil {
-		return "", err
-	}
-	return userinfo.Id, nil
 }
 
 func (o *OAuth) GetTokenFromCode(ctx context.Context, authCode string) (*oauth2.Token, error) {
